Up the test threshold slightly

+++ crates/viewer/re_viewer_context/src/test_context.rs
@@ -37,7 +37,7 @@ pub trait HarnessExt {
         broken_percent_threshold: f64,
     ) -> bool;
 
-    fn snapshot_with_broken_pixels(&mut self, name: &str, broken_pixels: i32);
+    fn snapshot_with_broken_pixels(&mut self, name: &str, broken_pixels: usize);
 }
 
 impl HarnessExt for egui_kittest::Harness<'_> {
@@ -102,7 +102,7 @@ impl HarnessExt for egui_kittest::Harness<'_> {
     }
 
     #[track_caller]
-    fn snapshot_with_broken_pixels(&mut self, name: &str, broken_pixels_threshold: i32) {
+    fn snapshot_with_broken_pixels(&mut self, name: &str, broken_pixels_threshold: usize) {
         match self.try_snapshot(name) {
             Ok(_) => {}
 
@@ -114,7 +114,7 @@ impl HarnessExt for egui_kittest::Harness<'_> {
                 } => {
                     re_log::debug!(num_broken_pixels, broken_pixels_threshold);
                     assert!(
-                        num_broken_pixels <= broken_pixels_threshold,
+                        num_broken_pixels as usize <= broken_pixels_threshold,
                         "{name} failed because {num_broken_pixels} > {broken_pixels_threshold}\n{diff_path:?}"
                     );
                 }
